added trial code for looking up star color based upon type.  updated orbit rendering code.  more font rendering work and testing.
the orbit rendering code works well.  this idea is to draw a curve of constant width on screen from the world position of the orbit.  there is a problem of artifacts IFF the orbit intersects the projection plane (through the screen).  my use case for rendering orbits will have the viewer looking "down" upon the plane in which the orbit(s) reside.